+++ search_params.go
@@ -63,6 +63,11 @@ type SearchParams struct {
 	// account instead of under the account of the owner of the configured
 	// Stripe key.
 	StripeAccount *string `form:"-"` // Passed as header
+
+	// StripeContext is used to set the Stripe-Context header on a request.
+	// The Stripe-Context header can be used to set the account with which
+	// the request is made. It is preferred to using StripeAccount in new code.
+	StripeContext *string `form:"-" json:"-"` // Passed as header
 }
 
 // AddExpand on the embedded SearchParams struct is deprecated
@@ -90,6 +95,18 @@ func (p *SearchParams) SetStripeAccount(val string) {
 	p.StripeAccount = &val
 }
 
+// SetStripeContext sets a value for the Stripe-Context header.
+func (p *SearchParams) SetStripeContext(val string) {
+	p.StripeContext = &val
+}
+
+// SetStripeContextFrom sets a value for the Stripe-Context header using a stripe.Context object.
+func (p *SearchParams) SetStripeContextFrom(val *Context) {
+	if val != nil {
+		p.StripeContext = val.StringPtr()
+	}
+}
+
 // ToParams converts a SearchParams to a Params by moving over any fields that
 // have valid targets in the new type. This is useful because fields in
 // Params can be injected directly into an http.Request while generally
@@ -98,6 +115,7 @@ func (p *SearchParams) ToParams() *Params {
 	return &Params{
 		Context:       p.Context,
 		StripeAccount: p.StripeAccount,
+		StripeContext: p.StripeContext,
 	}
 }